Worked on basic AR Functions.
I worked on the basic AR functionality to see what we can achieve.

I got the twerking zombie working (ARTtest_0.2.apk) and then a simple target that you can shoot (click on) that disapear. That means we have a script pretty much ready that allows palyers to click on shit in the AR world through their screens. (ARTest_0.25.apk)